Cambios de tipo y estilo

+++ src/herramientas/desgastador/DesgastadorLineal.java
@@ -2,15 +2,15 @@ package herramientas.desgastador;
 
 public class DesgastadorLineal implements Desgastador {
 
-    private float desgaste;
+    private int desgaste;
 
     public DesgastadorLineal(float fuerza, float i) {
-        desgaste = fuerza / i;
+        desgaste = (int) (fuerza / i);
     }
 
 
     @Override
-    public float calcularDesgaste() {
+    public int calcularDesgaste() {
         return desgaste;
     }
